[CAS] OnDiskTrieRawHashMap updates from upstream
Update downstream code after upstream review.
