+++ clang/test/CAS/daemon-cas-recovery.c
@@ -4,7 +4,7 @@
 
 /// Construct a malformed CAS to recovery from.
 // RUN: echo "abc" | llvm-cas --cas %t/cas --make-blob --data -
-// RUN: rm %t/cas/v1.1/v10.data
+// RUN: rm %t/cas/v1.1/v11.data
 // RUN: not llvm-cas --cas %t/cas --validate --check-hash
 
 // RUN: env LLVM_CACHE_CAS_PATH=%t/cas LLVM_CAS_FORCE_VALIDATION=1 %clang-cache \
