+++ llvm/include/llvm/CAS/OnDiskTrieRawHashMap.h
@@ -1,55 +1,41 @@
-//===- OnDiskHashMappedTrie.h -----------------------------------*- C++ -*-===//
+//===----------------------------------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
+//
+/// \file
+/// This file declares interface for OnDiskTrieRawHashMap, a thread-safe and
+/// (mostly) lock-free hash map stored as trie and backed by persistent files on
+/// disk.
+///
+//===----------------------------------------------------------------------===//
 
-#ifndef LLVM_CAS_ONDISKHASHMAPPEDTRIE_H
-#define LLVM_CAS_ONDISKHASHMAPPEDTRIE_H
+#ifndef LLVM_CAS_ONDISKTRIERAWHASHMAP_H
+#define LLVM_CAS_ONDISKTRIERAWHASHMAP_H
 
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/ADT/STLFunctionalExtras.h"
 #include "llvm/ADT/StringRef.h"
-#include "llvm/Config/llvm-config.h"
-#include "llvm/Support/Casting.h"
-#include "llvm/Support/Compiler.h"
-#include "llvm/Support/FileSystem.h"
-#include <atomic>
-#include <mutex>
+#include "llvm/CAS/FileOffset.h"
+#include "llvm/Support/Error.h"
 #include <optional>
 
 namespace llvm {
 
-class MemoryBuffer;
 class raw_ostream;
 
 namespace cas {
 
 namespace ondisk {
 class OnDiskCASLogger;
-}
-
-class FileOffset {
-public:
-  int64_t get() const { return Offset; }
-
-  explicit operator bool() const { return Offset; }
-
-  FileOffset() = default;
-  explicit FileOffset(int64_t Offset) : Offset(Offset) { assert(Offset >= 0); }
-
-private:
-  int64_t Offset = 0;
-};
-
-/// On-disk hash-mapped trie. Thread-safe / lock-free.
-///
-/// This is an on-disk, (mostly) thread-safe key-value store that is (mostly)
-/// lock-free. The keys are fixed length, and are expected to be binary hashes
-/// with a normal distribution.
+} // namespace ondisk
+/// OnDiskTrieRawHashMap is a persistent trie data structure used as hash maps.
+/// The keys are fixed length, and are expected to be binary hashes with a
+/// normal distribution.
 ///
 /// - Thread-safety is achieved through the use of atomics within a shared
 ///   memory mapping. Atomic access does not work on networked filesystems.
@@ -62,16 +48,10 @@ private:
 ///     - "index" stores the root trie and subtries.
 ///     - "data" stores (most of) the entries, like a bump-ptr-allocator.
 ///     - Large entries are stored externally in a file named by the key.
-/// - Code is system-dependent (Windows not yet implemented), and binary format
-///   itself is not portable. These are not artifacts that can/should be moved
-///   between different systems; they are only appropriate for local storage.
-///
-/// FIXME: Add support for storing top-level metadata or identifiers that can
-/// be created / read during initialization.
-///
-/// FIXME: Implement for Windows. See comment next to implementation of \a
-/// OnDiskHashMappedTrie::MappedFileInfo::open().
-class OnDiskHashMappedTrie {
+/// - Code is system-dependent and binary format itself is not portable. These
+///   are not artifacts that can/should be moved between different systems; they
+///   are only appropriate for local storage.
+class OnDiskTrieRawHashMap {
 public:
   LLVM_DUMP_METHOD void dump() const;
   void
@@ -79,6 +59,7 @@ public:
         function_ref<void(ArrayRef<char>)> PrintRecordData = nullptr) const;
 
 public:
+  /// Const value proxy to access the records stored in TrieRawHashMap.
   struct ConstValueProxy {
     ConstValueProxy() = default;
     ConstValueProxy(ArrayRef<uint8_t> Hash, ArrayRef<char> Data)
@@ -90,6 +71,7 @@ public:
     ArrayRef<char> Data;
   };
 
+  /// Value proxy to access the records stored in TrieRawHashMap.
   struct ValueProxy {
     operator ConstValueProxy() const { return ConstValueProxy(Hash, Data); }
 
@@ -107,7 +89,17 @@ public:
   Error validate(
       function_ref<Error(FileOffset, ConstValueProxy)> RecordVerifier) const;
 
+  /// Check the valid range of file offset for OnDiskTrieRawHashMap.
+  static bool validOffset(FileOffset Offset) {
+    return Offset.get() < (1LL << 48);
+  }
+
 public:
+  /// Template class to implement a `pointer` type into the trie data structure.
+  ///
+  /// It provides pointer-like operation, e.g., dereference to get underlying
+  /// data. It also reserves the top 16 bits of the pointer value, which can be
+  /// used to pack additional information if needed.
   template <class ProxyT> class PointerImpl {
   public:
     FileOffset getOffset() const {
@@ -128,24 +120,19 @@ public:
     PointerImpl() = default;
 
   protected:
-    PointerImpl(FileOffset Offset, ProxyT Value)
-        : PointerImpl(Value, Offset, /*IsValue=*/true) {}
-
-    PointerImpl(ProxyT Value, FileOffset Offset, bool IsValue)
+    PointerImpl(ProxyT Value, FileOffset Offset, bool IsValue = true)
         : Value(Value), OffsetLow32((uint64_t)Offset.get()),
           OffsetHigh16((uint64_t)Offset.get() >> 32), IsValue(IsValue) {
       if (IsValue)
-        checkOffset(Offset);
-    }
-
-    static void checkOffset(FileOffset Offset) {
-      assert(Offset.get() > 0);
-      assert((uint64_t)Offset.get() < (1LL << 48));
+        assert(validOffset(Offset));
     }
 
     ProxyT Value;
     uint32_t OffsetLow32 = 0;
     uint16_t OffsetHigh16 = 0;
+
+    // True if points to a value (not a "nullptr"). Use an extra field because
+    // 0 can be a valid offset.
     bool IsValue = false;
   };
 
@@ -156,7 +143,7 @@ public:
 
   private:
     friend class pointer;
-    friend class OnDiskHashMappedTrie;
+    friend class OnDiskTrieRawHashMap;
     using const_pointer::PointerImpl::PointerImpl;
   };
 
@@ -169,37 +156,18 @@ public:
     pointer() = default;
 
   private:
-    friend class OnDiskHashMappedTrie;
+    friend class OnDiskTrieRawHashMap;
     using pointer::PointerImpl::PointerImpl;
   };
 
-  pointer getMutablePointer(const_pointer CP) {
-    if (!CP)
-      return pointer();
-    ValueProxy V{CP->Hash, MutableArrayRef(const_cast<char *>(CP->Data.data()),
-                                           CP->Data.size())};
-    return pointer(CP.getOffset(), V);
-  }
-
+  /// Find the value from hash.
+  ///
+  /// \returns pointer to the value if exists, otherwise returns a non-value
+  /// pointer that evaluates to `false` when convert to boolean.
   const_pointer find(ArrayRef<uint8_t> Hash) const;
-  pointer find(ArrayRef<uint8_t> Hash) {
-    return getMutablePointer(
-        const_cast<const OnDiskHashMappedTrie *>(this)->find(Hash));
-  }
 
-  const_pointer recoverFromHashPointer(const uint8_t *HashBegin) const;
-  pointer recoverFromHashPointer(const uint8_t *HashBegin) {
-    return getMutablePointer(
-        const_cast<const OnDiskHashMappedTrie *>(this)->recoverFromHashPointer(
-            HashBegin));
-  }
-
-  const_pointer recoverFromFileOffset(FileOffset Offset) const;
-  pointer recoverFromFileOffset(FileOffset Offset) {
-    return getMutablePointer(
-        const_cast<const OnDiskHashMappedTrie *>(this)->recoverFromFileOffset(
-            Offset));
-  }
+  /// Helper function to recover a pointer into the trie from file offset.
+  Expected<const_pointer> recoverFromFileOffset(FileOffset Offset) const;
 
   using LazyInsertOnConstructCB =
       function_ref<void(FileOffset TentativeOffset, ValueProxy TentativeValue)>;
@@ -218,7 +186,7 @@ public:
   /// references.
   ///
   /// NOTE: Does *not* guarantee that \p OnConstruct is only called on success.
-  /// The in-memory \a HashMappedTrie uses LazyAtomicPointer to synchronize
+  /// The in-memory \a TrieRawHashMap uses LazyAtomicPointer to synchronize
   /// simultaneous writes, but that seems dangerous to use in a memory-mapped
   /// file in case a process crashes in the busy state.
   Expected<pointer> insertLazy(ArrayRef<uint8_t> Hash,
@@ -248,12 +216,7 @@ public:
   /// configure the trie, if it doesn't already exist.
   ///
   /// \pre NumHashBits is a multiple of 8 (byte-aligned).
-  ///
-  /// TODO: Expose the internal DatabaseFile abstraction and add support for
-  /// adding more tables to a single file.
-  ///
-  /// FIXME: Rename to getOrCreate().
-  static Expected<OnDiskHashMappedTrie>
+  static Expected<OnDiskTrieRawHashMap>
   create(const Twine &Path, const Twine &TrieName, size_t NumHashBits,
          uint64_t DataSize, uint64_t MaxFileSize,
          std::optional<uint64_t> NewFileInitialSize,
@@ -261,97 +224,17 @@ public:
          std::optional<size_t> NewTableNumRootBits = std::nullopt,
          std::optional<size_t> NewTableNumSubtrieBits = std::nullopt);
 
-  OnDiskHashMappedTrie(OnDiskHashMappedTrie &&RHS);
-  OnDiskHashMappedTrie &operator=(OnDiskHashMappedTrie &&RHS);
-  ~OnDiskHashMappedTrie();
+  OnDiskTrieRawHashMap(OnDiskTrieRawHashMap &&RHS);
+  OnDiskTrieRawHashMap &operator=(OnDiskTrieRawHashMap &&RHS);
+  ~OnDiskTrieRawHashMap();
 
 private:
   struct ImplType;
-  explicit OnDiskHashMappedTrie(std::unique_ptr<ImplType> Impl);
-  std::unique_ptr<ImplType> Impl;
-};
-
-/// Sink for data. Stores variable length data with 8-byte alignment. Does not
-/// track size of data, which is assumed to known from context, or embedded.
-/// Uses 0-padding but does not guarantee 0-termination.
-class OnDiskDataAllocator {
-public:
-  using ValueProxy = MutableArrayRef<char>;
-
-  /// An iterator-like return value for data insertion. Maybe it should be
-  /// called \c iterator, but it has no increment.
-  class pointer {
-  public:
-    FileOffset getOffset() const { return Offset; }
-    explicit operator bool() const { return bool(getOffset()); }
-    const ValueProxy &operator*() const {
-      assert(Offset && "Null dereference");
-      return Value;
-    }
-    const ValueProxy *operator->() const {
-      assert(Offset && "Null dereference");
-      return &Value;
-    }
-
-    pointer() = default;
-
-  private:
-    friend class OnDiskDataAllocator;
-    pointer(FileOffset Offset, ValueProxy Value)
-        : Offset(Offset), Value(Value) {}
-    FileOffset Offset;
-    ValueProxy Value;
-  };
-
-  // Look up the data stored at the given offset.
-  const char *beginData(FileOffset Offset) const;
-  char *beginData(FileOffset Offset) {
-    return const_cast<char *>(
-        const_cast<const OnDiskDataAllocator *>(this)->beginData(Offset));
-  }
-
-  Expected<pointer> allocate(size_t Size);
-  Expected<pointer> save(ArrayRef<char> Data) {
-    auto P = allocate(Data.size());
-    if (LLVM_UNLIKELY(!P))
-      return P.takeError();
-    llvm::copy(Data, (*P)->begin());
-    return P;
-  }
-  Expected<pointer> save(StringRef Data) {
-    return save(ArrayRef<char>(Data.begin(), Data.size()));
-  }
-
-  /// \returns the buffer that was allocated at \p create time, with size
-  /// \p UserHeaderSize.
-  MutableArrayRef<uint8_t> getUserHeader();
-
-  size_t size() const;
-  size_t capacity() const;
-
-  static Expected<OnDiskDataAllocator>
-  create(const Twine &Path, const Twine &TableName, uint64_t MaxFileSize,
-         std::optional<uint64_t> NewFileInitialSize,
-         uint32_t UserHeaderSize = 0,
-         std::shared_ptr<ondisk::OnDiskCASLogger> Logger = nullptr,
-         function_ref<void(void *)> UserHeaderInit = nullptr);
-
-  OnDiskDataAllocator(OnDiskDataAllocator &&RHS);
-  OnDiskDataAllocator &operator=(OnDiskDataAllocator &&RHS);
-
-  // No copy. Just call \a create() again.
-  OnDiskDataAllocator(const OnDiskDataAllocator &) = delete;
-  OnDiskDataAllocator &operator=(const OnDiskDataAllocator &) = delete;
-
-  ~OnDiskDataAllocator();
-
-private:
-  struct ImplType;
-  explicit OnDiskDataAllocator(std::unique_ptr<ImplType> Impl);
+  explicit OnDiskTrieRawHashMap(std::unique_ptr<ImplType> Impl);
   std::unique_ptr<ImplType> Impl;
 };
 
 } // namespace cas
 } // namespace llvm
 
-#endif // LLVM_CAS_ONDISKHASHMAPPEDTRIE_H
+#endif // LLVM_CAS_ONDISKTRIERAWHASHMAP_H
