+++ llvm/lib/CAS/OnDiskTrieRawHashMap.cpp
@@ -1,296 +1,56 @@
-//===- OnDiskHashMappedTrie.cpp -------------------------------------------===//
+//===----------------------------------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
+///
+/// \file Implements OnDiskTrieRawHashMap.
+///
+//===----------------------------------------------------------------------===//
 
-#include "llvm/CAS/OnDiskHashMappedTrie.h"
-#include "HashMappedTrieIndexGenerator.h"
-#include "llvm/ADT/ScopeExit.h"
+#include "llvm/CAS/OnDiskTrieRawHashMap.h"
+#include "DatabaseFile.h"
 #include "llvm/ADT/StringExtras.h"
-#include "llvm/ADT/StringMap.h"
+#include "llvm/ADT/TrieHashIndexGenerator.h"
 #include "llvm/CAS/MappedFileRegionArena.h"
 #include "llvm/CAS/OnDiskCASLogger.h"
-#include "llvm/Support/Compiler.h"
-#include "llvm/Support/Debug.h"
-#include "llvm/Support/Error.h"
-#include "llvm/Support/FileSystem.h"
-#include "llvm/Support/MemoryBuffer.h"
-#include "llvm/Support/Path.h"
+#include "llvm/Config/llvm-config.h"
 #include "llvm/Support/ThreadPool.h"
 #include "llvm/Support/Threading.h"
 #include "llvm/Support/raw_ostream.h"
-#include <mutex>
 
 using namespace llvm;
 using namespace llvm::cas;
-using ondisk::OnDiskCASLogger;
+using namespace llvm::cas::ondisk;
 
 #if LLVM_ENABLE_ONDISK_CAS
 
-static_assert(sizeof(size_t) == sizeof(uint64_t), "64-bit only");
-static_assert(sizeof(std::atomic<int64_t>) == sizeof(uint64_t),
-              "Requires lock-free 64-bit atomics");
-
 //===----------------------------------------------------------------------===//
-// Generic database data structures.
-//===----------------------------------------------------------------------===//
-namespace {
-using MappedFileRegion = MappedFileRegionArena::RegionT;
-
-/// Generic handle for a table.
-///
-/// Probably we want some table kinds for pointing at multiple tables.
-/// - Probably a tree or trie type makes sense.
-/// - Or a deque. Linear search is okay as long as there aren't many tables in
-///   a file.
-///
-/// Generic table header layout:
-/// - 2-bytes: TableKind
-/// - 2-bytes: TableNameSize
-/// - 4-bytes: TableNameRelOffset (relative to header)
-class TableHandle {
-public:
-  enum class TableKind : uint16_t {
-    HashMappedTrie = 1,
-    DataAllocator = 2,
-  };
-  struct Header {
-    TableKind Kind;
-    uint16_t NameSize;
-    int32_t NameRelOffset; // Relative to Header.
-  };
-
-  explicit operator bool() const { return H; }
-  const Header &getHeader() const { return *H; }
-  MappedFileRegion &getRegion() const { return *Region; }
-
-  template <class T> static void check() {
-    static_assert(
-        std::is_same<decltype(T::Header::GenericHeader), Header>::value,
-        "T::GenericHeader should be of type TableHandle::Header");
-    static_assert(offsetof(typename T::Header, GenericHeader) == 0,
-                  "T::GenericHeader must be the head of T::Header");
-  }
-  template <class T> bool is() const { return T::Kind == H->Kind; }
-  template <class T> T dyn_cast() const {
-    check<T>();
-    if (is<T>())
-      return T(*Region, *reinterpret_cast<typename T::Header *>(H));
-    return T();
-  }
-  template <class T> T cast() const {
-    assert(is<T>());
-    return dyn_cast<T>();
-  }
-
-  StringRef getName() const {
-    auto *Begin = reinterpret_cast<const char *>(H) + H->NameRelOffset;
-    return StringRef(Begin, H->NameSize);
-  }
-
-  TableHandle() = default;
-  TableHandle(MappedFileRegion &Region, Header &H) : Region(&Region), H(&H) {}
-  TableHandle(MappedFileRegion &Region, intptr_t HeaderOffset)
-      : TableHandle(Region,
-                    *reinterpret_cast<Header *>(Region.data() + HeaderOffset)) {
-  }
-
-private:
-  MappedFileRegion *Region = nullptr;
-  Header *H = nullptr;
-};
-
-/// Encapsulate a database file, which:
-/// - Sets/checks magic.
-/// - Sets/checks version.
-/// - Points at an arbitrary root table (can be changed later using a lock-free
-///   algorithm).
-/// - Sets up a BumpPtr for allocation.
-///
-/// Top-level layout:
-/// - 8-bytes: Magic
-/// - 8-bytes: Version
-/// - 8-bytes: RootTable (16-bits: Kind; 48-bits: Offset)
-/// - 8-bytes: BumpPtr
-class DatabaseFile {
-public:
-  static constexpr uint64_t getMagic() { return 0x00FFDA7ABA53FF00ULL; }
-  static constexpr uint64_t getVersion() { return 1ULL; }
-  struct Header {
-    uint64_t Magic;
-    uint64_t Version;
-    std::atomic<int64_t> RootTableOffset;
-  };
-
-  const Header &getHeader() { return *H; }
-  MappedFileRegionArena &getAlloc() { return Alloc; }
-  MappedFileRegion &getRegion() { return Alloc.getRegion(); }
-
-  /// Add a table.
-  ///
-  /// TODO: Allow lazy construction via getOrCreate()-style API.
-  void addTable(TableHandle Table);
-
-  /// Find a table. May return null.
-  std::optional<TableHandle> findTable(StringRef Name);
-
-  static Expected<DatabaseFile>
-  create(const Twine &Path, uint64_t Capacity,
-         std::shared_ptr<OnDiskCASLogger> Logger,
-         function_ref<Error(DatabaseFile &)> NewDBConstructor);
-
-  size_t size() const { return Alloc.size(); }
-
-private:
-  static Expected<DatabaseFile>
-  get(std::unique_ptr<MappedFileRegionArena> Alloc) {
-    if (Error E = validate(Alloc->getRegion()))
-      return std::move(E);
-    return DatabaseFile(std::move(Alloc));
-  }
-
-  static Error validate(MappedFileRegion &Region);
-
-  DatabaseFile(MappedFileRegionArena &Alloc)
-      : H(reinterpret_cast<Header *>(Alloc.data())), Alloc(Alloc) {}
-  DatabaseFile(std::unique_ptr<MappedFileRegionArena> Alloc)
-      : DatabaseFile(*Alloc) {
-    OwnedAlloc = std::move(Alloc);
-  }
-
-  Header *H = nullptr;
-  MappedFileRegionArena &Alloc;
-  std::unique_ptr<MappedFileRegionArena> OwnedAlloc;
-};
-
-} // end anonymous namespace
-
-static Error createTableConfigError(std::errc ErrC, StringRef Path,
-                                    StringRef TableName, const Twine &Msg) {
-  return createStringError(make_error_code(ErrC),
-                           Path + "[" + TableName + "]: " + Msg);
-}
-
-Expected<DatabaseFile>
-DatabaseFile::create(const Twine &Path, uint64_t Capacity,
-                     std::shared_ptr<OnDiskCASLogger> Logger,
-                     function_ref<Error(DatabaseFile &)> NewDBConstructor) {
-  // Constructor for if the file doesn't exist.
-  auto NewFileConstructor = [&](MappedFileRegionArena &Alloc) -> Error {
-    if (Alloc.capacity() <
-        sizeof(Header) + sizeof(MappedFileRegionArena::Header))
-      return createTableConfigError(std::errc::argument_out_of_domain,
-                                    Path.str(), "datafile",
-                                    "Allocator too small for header");
-    (void)new (Alloc.data()) Header{getMagic(), getVersion(), {0}};
-    DatabaseFile DB(Alloc);
-    return NewDBConstructor(DB);
-  };
-
-  // Get or create the file.
-  MappedFileRegionArena Alloc;
-  if (Error E =
-          MappedFileRegionArena::create(Path, Capacity, sizeof(Header),
-                                        std::move(Logger), NewFileConstructor)
-              .moveInto(Alloc))
-    return std::move(E);
-
-  return DatabaseFile::get(
-      std::make_unique<MappedFileRegionArena>(std::move(Alloc)));
-}
-
-void DatabaseFile::addTable(TableHandle Table) {
-  assert(Table);
-  assert(&Table.getRegion() == &getRegion());
-  int64_t ExistingRootOffset = 0;
-  const int64_t NewOffset =
-      reinterpret_cast<const char *>(&Table.getHeader()) - getRegion().data();
-  if (H->RootTableOffset.compare_exchange_strong(ExistingRootOffset, NewOffset))
-    return;
-
-  // Silently ignore attempts to set the root to itself.
-  if (ExistingRootOffset == NewOffset)
-    return;
-
-  // FIXME: Fix the API so that having the same name is not an error. Instead,
-  // the colliding table should just be used as-is and the client can decide
-  // what to do with the new one.
-  //
-  // TODO: Add support for creating a chain or tree of tables (more than one at
-  // all!) to avoid this error.
-  TableHandle Root(getRegion(), ExistingRootOffset);
-  if (Root.getName() == Table.getName())
-    report_fatal_error(
-        createStringError(make_error_code(std::errc::not_supported),
-                          "table name collision '" + Table.getName() + "'"));
-  else
-    report_fatal_error(
-        createStringError(make_error_code(std::errc::not_supported),
-                          "cannot add new table '" + Table.getName() +
-                              "'"
-                              " to existing root '" +
-                              Root.getName() + "'"));
-}
-
-std::optional<TableHandle> DatabaseFile::findTable(StringRef Name) {
-  int64_t RootTableOffset = H->RootTableOffset.load();
-  if (!RootTableOffset)
-    return std::nullopt;
-
-  TableHandle Root(getRegion(), RootTableOffset);
-  if (Root.getName() == Name)
-    return Root;
-
-  // TODO: Once multiple tables are supported, need to walk to find them.
-  return std::nullopt;
-}
-
-Error DatabaseFile::validate(MappedFileRegion &Region) {
-  if (Region.size() < sizeof(Header))
-    return createStringError(std::errc::invalid_argument,
-                             "database: missing header");
-
-  // Check the magic and version.
-  auto *H = reinterpret_cast<Header *>(Region.data());
-  if (H->Magic != getMagic())
-    return createStringError(std::errc::invalid_argument,
-                             "database: bad magic");
-  if (H->Version != getVersion())
-    return createStringError(std::errc::invalid_argument,
-                             "database: wrong version");
-
-  auto *MFH = reinterpret_cast<MappedFileRegionArena::Header *>(Region.data() +
-                                                                sizeof(Header));
-  // Check the bump-ptr, which should point past the header.
-  if (MFH->BumpPtr.load() < (int64_t)sizeof(Header))
-    return createStringError(std::errc::invalid_argument,
-                             "database: corrupt bump-ptr");
-
-  return Error::success();
-}
-
-//===----------------------------------------------------------------------===//
-// HashMappedTrie data structures.
+// TrieRawHashMap data structures.
 //===----------------------------------------------------------------------===//
 
 namespace {
 
 class SubtrieHandle;
+class TrieRawHashMapHandle;
 class TrieVisitor;
+
+/// A value stored in the slots inside a SubTrie. A stored value can either be a
+/// subtrie (encoded after negation) which is the file offset to another
+/// subtrie, or it can be a fileset to a DataRecord.
 class SubtrieSlotValue {
 public:
   explicit operator bool() const { return !isEmpty(); }
   bool isEmpty() const { return !Offset; }
   bool isData() const { return Offset > 0; }
   bool isSubtrie() const { return Offset < 0; }
-  int64_t asData() const {
+  uint64_t asData() const {
     assert(isData());
     return Offset;
   }
-  int64_t asSubtrie() const {
+  uint64_t asSubtrie() const {
     assert(isSubtrie());
     return -Offset;
   }
@@ -328,8 +88,6 @@ private:
   explicit SubtrieSlotValue(int64_t Offset) : Offset(Offset) {}
   int64_t Offset = 0;
 };
-
-class HashMappedTrieHandle;
 
 /// Subtrie layout:
 /// - 2-bytes: StartBit
@@ -451,11 +209,11 @@ private:
   }
 };
 
-/// Handle for a HashMappedTrie table.
+/// Handle for a TrieRawHashMap table.
 ///
-/// HashMappedTrie table layout:
+/// TrieRawHashMap table layout:
 /// - [8-bytes: Generic table header]
-/// - 1-byte: NumSubtrieBits
+/// - 1-byte:  NumSubtrieBits
 /// - 1-byte:  Flags (not used yet)
 /// - 2-bytes: NumHashBits
 /// - 4-bytes: RecordDataSize (in bytes)
@@ -464,17 +222,17 @@ private:
 /// - <name> '\0'
 ///
 /// Record layout:
-/// - <data>
 /// - <hash>
-class HashMappedTrieHandle {
+/// - <data>
+class TrieRawHashMapHandle {
 public:
   static constexpr TableHandle::TableKind Kind =
-      TableHandle::TableKind::HashMappedTrie;
+      TableHandle::TableKind::TrieRawHashMap;
 
   struct Header {
     TableHandle::Header GenericHeader;
     uint8_t NumSubtrieBits;
-    uint8_t Flags; // None used yet.
+    uint8_t Flags; ///< None used yet.
     uint16_t NumHashBits;
     uint32_t RecordDataSize;
     std::atomic<int64_t> RootTrieOffset;
@@ -488,7 +246,7 @@ public:
   }
 
   struct RecordData {
-    OnDiskHashMappedTrie::ValueProxy Proxy;
+    OnDiskTrieRawHashMap::ValueProxy Proxy;
     SubtrieSlotValue Offset;
     FileOffset getFileOffset() const { return Offset.asDataFileOffset(); }
   };
@@ -535,14 +293,15 @@ public:
     return getRecordSize(H->RecordDataSize, H->NumHashBits);
   }
 
-  IndexGenerator getIndexGen(SubtrieHandle Root, ArrayRef<uint8_t> Hash) {
+  TrieHashIndexGenerator getIndexGen(SubtrieHandle Root,
+                                     ArrayRef<uint8_t> Hash) {
     assert(Root.getStartBit() == 0);
     assert(getNumHashBytes() == Hash.size());
     assert(getNumHashBits() == Hash.size() * 8);
-    return IndexGenerator{Root.getNumBits(), getNumSubtrieBits(), Hash};
+    return TrieHashIndexGenerator{Root.getNumBits(), getNumSubtrieBits(), Hash};
   }
 
-  static Expected<HashMappedTrieHandle>
+  static Expected<TrieRawHashMapHandle>
   create(MappedFileRegionArena &Alloc, StringRef Name,
          std::optional<uint64_t> NumRootBits, uint64_t NumSubtrieBits,
          uint64_t NumHashBits, uint64_t RecordDataSize,
@@ -553,16 +312,15 @@ public:
         function_ref<void(ArrayRef<char>)> PrintRecordData = nullptr) const;
 
   Error validate(
-      function_ref<Error(FileOffset, OnDiskHashMappedTrie::ConstValueProxy)>
+      function_ref<Error(FileOffset, OnDiskTrieRawHashMap::ConstValueProxy)>
           RecordVerifier) const;
-
-  HashMappedTrieHandle() = default;
-  HashMappedTrieHandle(MappedFileRegion &Region, Header &H,
+  TrieRawHashMapHandle() = default;
+  TrieRawHashMapHandle(MappedFileRegion &Region, Header &H,
                        std::shared_ptr<OnDiskCASLogger> Logger = nullptr)
       : Region(&Region), H(&H), Logger(std::move(Logger)) {}
-  HashMappedTrieHandle(MappedFileRegion &Region, intptr_t HeaderOffset,
+  TrieRawHashMapHandle(MappedFileRegion &Region, intptr_t HeaderOffset,
                        std::shared_ptr<OnDiskCASLogger> Logger = nullptr)
-      : HashMappedTrieHandle(
+      : TrieRawHashMapHandle(
             Region, *reinterpret_cast<Header *>(Region.data() + HeaderOffset),
             std::move(Logger)) {}
 
@@ -579,18 +337,18 @@ private:
 
 } // end anonymous namespace
 
-struct OnDiskHashMappedTrie::ImplType {
+struct OnDiskTrieRawHashMap::ImplType {
   DatabaseFile File;
-  HashMappedTrieHandle Trie;
+  TrieRawHashMapHandle Trie;
 };
 
 Expected<SubtrieHandle> SubtrieHandle::create(MappedFileRegionArena &Alloc,
                                               uint32_t StartBit,
                                               uint32_t NumBits,
                                               OnDiskCASLogger *Logger) {
-  assert(StartBit <= HashMappedTrieHandle::MaxNumHashBits);
+  assert(StartBit <= TrieRawHashMapHandle::MaxNumHashBits);
   assert(NumBits <= UINT8_MAX);
-  assert(NumBits <= HashMappedTrieHandle::MaxNumRootBits);
+  assert(NumBits <= TrieRawHashMapHandle::MaxNumRootBits);
 
   auto Mem = Alloc.allocate(getSize(NumBits));
   if (LLVM_UNLIKELY(!Mem))
@@ -608,7 +366,7 @@ Expected<SubtrieHandle> SubtrieHandle::create(MappedFileRegionArena &Alloc,
   return S;
 }
 
-SubtrieHandle HashMappedTrieHandle::getRoot() const {
+SubtrieHandle TrieRawHashMapHandle::getRoot() const {
   if (int64_t Root = H->RootTrieOffset)
     return SubtrieHandle(getRegion(), SubtrieSlotValue::getSubtrieOffset(Root),
                          Logger.get());
@@ -616,7 +374,7 @@ SubtrieHandle HashMappedTrieHandle::getRoot() const {
 }
 
 Expected<SubtrieHandle>
-HashMappedTrieHandle::getOrCreateRoot(MappedFileRegionArena &Alloc) {
+TrieRawHashMapHandle::getOrCreateRoot(MappedFileRegionArena &Alloc) {
   assert(&Alloc.getRegion() == &getRegion());
   if (SubtrieHandle Root = getRoot())
     return Root;
@@ -639,8 +397,8 @@ HashMappedTrieHandle::getOrCreateRoot(MappedFileRegionArena &Alloc) {
                        Logger.get());
 }
 
-Expected<HashMappedTrieHandle>
-HashMappedTrieHandle::create(MappedFileRegionArena &Alloc, StringRef Name,
+Expected<TrieRawHashMapHandle>
+TrieRawHashMapHandle::create(MappedFileRegionArena &Alloc, StringRef Name,
                              std::optional<uint64_t> NumRootBits,
                              uint64_t NumSubtrieBits, uint64_t NumHashBits,
                              uint64_t RecordDataSize,
@@ -656,7 +414,7 @@ HashMappedTrieHandle::create(MappedFileRegionArena &Alloc, StringRef Name,
   assert(NumHashBits <= UINT16_MAX && "Expected valid hash size");
   assert(RecordDataSize <= UINT32_MAX && "Expected smaller table name");
   auto *H = new (Alloc.getRegion().data() + *Offset)
-      Header{{TableHandle::TableKind::HashMappedTrie, (uint16_t)Name.size(),
+      Header{{TableHandle::TableKind::TrieRawHashMap, (uint16_t)Name.size(),
               (uint32_t)sizeof(Header)},
              (uint8_t)NumSubtrieBits,
              /*Flags=*/0,
@@ -669,7 +427,7 @@ HashMappedTrieHandle::create(MappedFileRegionArena &Alloc, StringRef Name,
   NameStorage[Name.size()] = 0;
 
   // Construct a root trie, if requested.
-  HashMappedTrieHandle Trie(Alloc.getRegion(), *H, Logger);
+  TrieRawHashMapHandle Trie(Alloc.getRegion(), *H, Logger);
   auto Sub = SubtrieHandle::create(Alloc, 0, *NumRootBits, Logger.get());
   if (LLVM_UNLIKELY(!Sub))
     return Sub.takeError();
@@ -678,18 +436,18 @@ HashMappedTrieHandle::create(MappedFileRegionArena &Alloc, StringRef Name,
   return Trie;
 }
 
-HashMappedTrieHandle::RecordData
-HashMappedTrieHandle::getRecord(SubtrieSlotValue Offset) {
+TrieRawHashMapHandle::RecordData
+TrieRawHashMapHandle::getRecord(SubtrieSlotValue Offset) {
   char *Begin = Region->data() + Offset.asData();
-  OnDiskHashMappedTrie::ValueProxy Proxy;
+  OnDiskTrieRawHashMap::ValueProxy Proxy;
   Proxy.Data = MutableArrayRef(Begin, getRecordDataSize());
   Proxy.Hash = ArrayRef(reinterpret_cast<const uint8_t *>(Proxy.Data.end()),
-                            getNumHashBytes());
+                        getNumHashBytes());
   return RecordData{Proxy, Offset};
 }
 
-Expected<HashMappedTrieHandle::RecordData>
-HashMappedTrieHandle::createRecord(MappedFileRegionArena &Alloc,
+Expected<TrieRawHashMapHandle::RecordData>
+TrieRawHashMapHandle::createRecord(MappedFileRegionArena &Alloc,
                                    ArrayRef<uint8_t> Hash) {
   assert(&Alloc.getRegion() == Region);
   assert(Hash.size() == getNumHashBytes());
@@ -707,57 +465,40 @@ HashMappedTrieHandle::createRecord(MappedFileRegionArena &Alloc,
   return Record;
 }
 
-OnDiskHashMappedTrie::const_pointer
-OnDiskHashMappedTrie::recoverFromHashPointer(
-    const uint8_t *HashBeginPtr) const {
-  // Record hashes occur immediately after data. Compute the beginning of the
-  // record and check for overflow.
-  const uintptr_t HashBegin = reinterpret_cast<uintptr_t>(HashBeginPtr);
-  const uintptr_t RecordBegin = HashBegin - Impl->Trie.getRecordSize();
-  if (HashBegin < RecordBegin)
-    return const_pointer();
-
-  // Check that it'll be a positive offset.
-  const uintptr_t FileBegin =
-      reinterpret_cast<uintptr_t>(Impl->File.getRegion().data());
-  if (RecordBegin < FileBegin)
-    return const_pointer();
-
-  // Good enough to form an offset. Continue checking there.
-  return recoverFromFileOffset(FileOffset(RecordBegin - FileBegin));
-}
-
-OnDiskHashMappedTrie::const_pointer
-OnDiskHashMappedTrie::recoverFromFileOffset(FileOffset Offset) const {
+Expected<OnDiskTrieRawHashMap::const_pointer>
+OnDiskTrieRawHashMap::recoverFromFileOffset(FileOffset Offset) const {
   // Check alignment.
   if (!isAligned(MappedFileRegionArena::getAlign(), Offset.get()))
-    return const_pointer();
+    return createStringError(make_error_code(std::errc::protocol_error),
+                             "unaligned file offset at 0x" +
+                                 utohexstr(Offset.get(), /*LowerCase=*/true));
 
   // Check bounds.
   //
   // Note: There's no potential overflow when using \c uint64_t because Offset
-  // is in \c [0,INT64_MAX] and the record size is in \c [0,UINT32_MAX].
-  assert(Offset.get() >= 0 && "Expected FileOffset constructor guarantee this");
-  if ((uint64_t)Offset.get() + Impl->Trie.getRecordSize() >
-      Impl->File.getAlloc().size())
-    return const_pointer();
+  // is in valid offset range and the record size is in \c [0,UINT32_MAX].
+  if (!validOffset(Offset) ||
+      Offset.get() + Impl->Trie.getRecordSize() > Impl->File.getAlloc().size())
+    return createStringError(make_error_code(std::errc::protocol_error),
+                             "file offset too large: 0x" +
+                                 utohexstr(Offset.get(), /*LowerCase=*/true));
 
   // Looks okay...
-  HashMappedTrieHandle::RecordData D =
+  TrieRawHashMapHandle::RecordData D =
       Impl->Trie.getRecord(SubtrieSlotValue::getDataOffset(Offset));
-  return const_pointer(D.getFileOffset(), D.Proxy);
+  return const_pointer(D.Proxy, D.getFileOffset());
 }
 
-OnDiskHashMappedTrie::const_pointer
-OnDiskHashMappedTrie::find(ArrayRef<uint8_t> Hash) const {
-  HashMappedTrieHandle Trie = Impl->Trie;
+OnDiskTrieRawHashMap::const_pointer
+OnDiskTrieRawHashMap::find(ArrayRef<uint8_t> Hash) const {
+  TrieRawHashMapHandle Trie = Impl->Trie;
   assert(Hash.size() == Trie.getNumHashBytes() && "Invalid hash");
 
   SubtrieHandle S = Trie.getRoot();
   if (!S)
     return const_pointer();
 
-  IndexGenerator IndexGen = Trie.getIndexGen(S, Hash);
+  TrieHashIndexGenerator IndexGen = Trie.getIndexGen(S, Hash);
   size_t Index = IndexGen.next();
   for (;;) {
     // Try to set the content.
@@ -767,8 +508,8 @@ OnDiskHashMappedTrie::find(ArrayRef<uint8_t> Hash) const {
 
     // Check for an exact match.
     if (V.isData()) {
-      HashMappedTrieHandle::RecordData D = Trie.getRecord(V);
-      return D.Proxy.Hash == Hash ? const_pointer(D.getFileOffset(), D.Proxy)
+      TrieRawHashMapHandle::RecordData D = Trie.getRecord(V);
+      return D.Proxy.Hash == Hash ? const_pointer(D.Proxy, D.getFileOffset())
                                   : const_pointer();
     }
 
@@ -787,11 +528,11 @@ void SubtrieHandle::reinitialize(uint32_t StartBit, uint32_t NumBits) {
   H->NumBits = NumBits;
 }
 
-Expected<OnDiskHashMappedTrie::pointer>
-OnDiskHashMappedTrie::insertLazy(ArrayRef<uint8_t> Hash,
+Expected<OnDiskTrieRawHashMap::pointer>
+OnDiskTrieRawHashMap::insertLazy(ArrayRef<uint8_t> Hash,
                                  LazyInsertOnConstructCB OnConstruct,
                                  LazyInsertOnLeakCB OnLeak) {
-  HashMappedTrieHandle Trie = Impl->Trie;
+  TrieRawHashMapHandle Trie = Impl->Trie;
   assert(Hash.size() == Trie.getNumHashBytes() && "Invalid hash");
 
   MappedFileRegionArena &Alloc = Impl->File.getAlloc();
@@ -800,22 +541,11 @@ OnDiskHashMappedTrie::insertLazy(ArrayRef<uint8_t> Hash,
   if (LLVM_UNLIKELY(Err))
     return std::move(Err);
 
-  IndexGenerator IndexGen = Trie.getIndexGen(*S, Hash);
+  TrieHashIndexGenerator IndexGen = Trie.getIndexGen(*S, Hash);
   size_t Index = IndexGen.next();
 
-  // FIXME: Add non-assertion based checks for data corruption that would
-  // otherwise cause infinite loops in release builds, instead calling
-  // report_fatal_error().
-  //
-  // Two loops are possible:
-  // - All bits used up in the IndexGenerator because subtries are somehow
-  //   linked in a cycle. Could confirm that each subtrie's start-bit
-  //   follows from the start-bit and num-bits of its parent. Could also check
-  //   that the generator doesn't run out of bits.
-  // - Existing data matches tail of Hash but not the head (stored in an
-  //   invalid spot). Probably a cheap way to check this too, but needs
-  //   thought.
-  std::optional<HashMappedTrieHandle::RecordData> NewRecord;
+  // Walk through the hash bytes and insert into correct trie position.
+  std::optional<TrieRawHashMapHandle::RecordData> NewRecord;
   SubtrieHandle UnusedSubtrie;
   for (;;) {
     SubtrieSlotValue Existing = S->load(Index);
@@ -831,7 +561,7 @@ OnDiskHashMappedTrie::insertLazy(ArrayRef<uint8_t> Hash,
       }
 
       if (S->compare_exchange_strong(Index, Existing, NewRecord->Offset))
-        return pointer(NewRecord->Offset.asDataFileOffset(), NewRecord->Proxy);
+        return pointer(NewRecord->Proxy, NewRecord->Offset.asDataFileOffset());
 
       // Race means that Existing is no longer empty; fall through...
     }
@@ -843,13 +573,13 @@ OnDiskHashMappedTrie::insertLazy(ArrayRef<uint8_t> Hash,
     }
 
     // Check for an exact match.
-    HashMappedTrieHandle::RecordData ExistingRecord = Trie.getRecord(Existing);
+    TrieRawHashMapHandle::RecordData ExistingRecord = Trie.getRecord(Existing);
     if (ExistingRecord.Proxy.Hash == Hash) {
       if (NewRecord && OnLeak)
         OnLeak(NewRecord->Offset.asDataFileOffset(), NewRecord->Proxy,
                ExistingRecord.Offset.asDataFileOffset(), ExistingRecord.Proxy);
-      return pointer(ExistingRecord.Offset.asDataFileOffset(),
-                     ExistingRecord.Proxy);
+      return pointer(ExistingRecord.Proxy,
+                     ExistingRecord.Offset.asDataFileOffset());
     }
 
     // Sink the existing content as long as the indexes match.
@@ -907,23 +637,17 @@ Expected<SubtrieHandle> SubtrieHandle::sink(size_t I, SubtrieSlotValue V,
   return SubtrieHandle(Alloc.getRegion(), V, Logger);
 }
 
-void OnDiskHashMappedTrie::print(
+void OnDiskTrieRawHashMap::print(
     raw_ostream &OS, function_ref<void(ArrayRef<char>)> PrintRecordData) const {
   Impl->Trie.print(OS, PrintRecordData);
 }
 
-Error OnDiskHashMappedTrie::validate(
+Error OnDiskTrieRawHashMap::validate(
     function_ref<Error(FileOffset, ConstValueProxy)> RecordVerifier) const {
   return Impl->Trie.validate(RecordVerifier);
 }
 
-static void printHexDigit(raw_ostream &OS, uint8_t Digit) {
-  if (Digit < 10)
-    OS << char(Digit + '0');
-  else
-    OS << char(Digit - 10 + 'a');
-}
-
+// Helper function that prints hexdigit and have a sub-byte starting position.
 static void printHexDigits(raw_ostream &OS, ArrayRef<uint8_t> Bytes,
                            size_t StartBit, size_t NumBits) {
   assert(StartBit % 4 == 0);
@@ -931,7 +655,7 @@ static void printHexDigits(raw_ostream &OS, ArrayRef<uint8_t> Bytes,
   for (size_t I = StartBit, E = StartBit + NumBits; I != E; I += 4) {
     uint8_t HexPair = Bytes[I / 8];
     uint8_t HexDigit = I % 8 == 0 ? HexPair >> 4 : HexPair & 0xf;
-    printHexDigit(OS, HexDigit);
+    OS << hexdigit(HexDigit, /*LowerCase=*/true);
   }
 }
 
@@ -980,14 +704,14 @@ static void printPrefix(raw_ostream &OS, StringRef Prefix) {
     bool ErrorParsingBinary = Prefix.take_front(4).getAsInteger(2, Digit);
     assert(!ErrorParsingBinary);
     (void)ErrorParsingBinary;
-    printHexDigit(OS, Digit);
+    OS << hexdigit(Digit, /*LowerCase=*/true);
     Prefix = Prefix.drop_front(4);
   }
   if (!Prefix.empty())
     OS << "[" << Prefix << "]";
 }
 
-LLVM_DUMP_METHOD void OnDiskHashMappedTrie::dump() const { print(dbgs()); }
+LLVM_DUMP_METHOD void OnDiskTrieRawHashMap::dump() const { print(dbgs()); }
 
 static Expected<size_t> checkParameter(StringRef Label, size_t Max,
                                        std::optional<size_t> Value,
@@ -1005,23 +729,13 @@ static Expected<size_t> checkParameter(StringRef Label, size_t Max,
       "invalid " + Label + ": " + Twine(*Value) + " (max: " + Twine(Max) + ")");
 }
 
-static Error checkTable(StringRef Label, size_t Expected, size_t Observed,
-                        StringRef Path, StringRef TrieName) {
-  if (Expected == Observed)
-    return Error::success();
-  return createTableConfigError(std::errc::invalid_argument, Path, TrieName,
-                                "mismatched " + Label +
-                                    " (expected: " + Twine(Expected) +
-                                    ", observed: " + Twine(Observed) + ")");
-}
-
-size_t OnDiskHashMappedTrie::size() const { return Impl->File.size(); }
-size_t OnDiskHashMappedTrie::capacity() const {
+size_t OnDiskTrieRawHashMap::size() const { return Impl->File.size(); }
+size_t OnDiskTrieRawHashMap::capacity() const {
   return Impl->File.getRegion().size();
 }
 
-Expected<OnDiskHashMappedTrie>
-OnDiskHashMappedTrie::create(const Twine &PathTwine, const Twine &TrieNameTwine,
+Expected<OnDiskTrieRawHashMap>
+OnDiskTrieRawHashMap::create(const Twine &PathTwine, const Twine &TrieNameTwine,
                              size_t NumHashBits, uint64_t DataSize,
                              uint64_t MaxFileSize,
                              std::optional<uint64_t> NewFileInitialSize,
@@ -1038,14 +752,14 @@ OnDiskHashMappedTrie::create(const Twine &PathTwine, const Twine &TrieNameTwine,
 
   size_t NumRootBits;
   if (Error E = checkParameter(
-                    "root bits", HashMappedTrieHandle::MaxNumRootBits,
+                    "root bits", TrieRawHashMapHandle::MaxNumRootBits,
                     NewTableNumRootBits, DefaultNumRootBits, Path, TrieName)
                     .moveInto(NumRootBits))
     return std::move(E);
 
   size_t NumSubtrieBits;
   if (Error E = checkParameter("subtrie bits",
-                               HashMappedTrieHandle::MaxNumSubtrieBits,
+                               TrieRawHashMapHandle::MaxNumSubtrieBits,
                                NewTableNumSubtrieBits, DefaultNumSubtrieBits,
                                Path, TrieName)
                     .moveInto(NumSubtrieBits))
@@ -1053,7 +767,7 @@ OnDiskHashMappedTrie::create(const Twine &PathTwine, const Twine &TrieNameTwine,
 
   size_t NumHashBytes = NumHashBits >> 3;
   if (Error E =
-          checkParameter("hash size", HashMappedTrieHandle::MaxNumHashBits,
+          checkParameter("hash size", TrieRawHashMapHandle::MaxNumHashBits,
                          NumHashBits, std::nullopt, Path, TrieName)
               .takeError())
     return std::move(E);
@@ -1066,14 +780,13 @@ OnDiskHashMappedTrie::create(const Twine &PathTwine, const Twine &TrieNameTwine,
 
   // Constructor for if the file doesn't exist.
   auto NewDBConstructor = [&](DatabaseFile &DB) -> Error {
-    auto Trie = HashMappedTrieHandle::create(DB.getAlloc(), TrieName,
+    auto Trie = TrieRawHashMapHandle::create(DB.getAlloc(), TrieName,
                                              NumRootBits, NumSubtrieBits,
                                              NumHashBits, DataSize, Logger);
     if (LLVM_UNLIKELY(!Trie))
       return Trie.takeError();
 
-    DB.addTable(*Trie);
-    return Error::success();
+    return DB.addTable(*Trie);
   };
 
   // Get or create the file.
@@ -1083,16 +796,14 @@ OnDiskHashMappedTrie::create(const Twine &PathTwine, const Twine &TrieNameTwine,
     return File.takeError();
 
   // Find the trie and validate it.
-  //
-  // TODO: Add support for creating/adding a table to an existing file.
   std::optional<TableHandle> Table = File->findTable(TrieName);
   if (!Table)
     return createTableConfigError(std::errc::argument_out_of_domain, Path,
                                   TrieName, "table not found");
-  if (Error E = checkTable("table kind", (size_t)HashMappedTrieHandle::Kind,
+  if (Error E = checkTable("table kind", (size_t)TrieRawHashMapHandle::Kind,
                            (size_t)Table->getHeader().Kind, Path, TrieName))
     return std::move(E);
-  auto Trie = Table->cast<HashMappedTrieHandle>();
+  auto Trie = Table->cast<TrieRawHashMapHandle>();
   Trie.setLogger(Logger);
   assert(Trie && "Already checked the kind");
 
@@ -1111,8 +822,8 @@ OnDiskHashMappedTrie::create(const Twine &PathTwine, const Twine &TrieNameTwine,
                                   "unsupported flags: " + Twine(Flags));
 
   // Success.
-  OnDiskHashMappedTrie::ImplType Impl{DatabaseFile(std::move(*File)), Trie};
-  return OnDiskHashMappedTrie(std::make_unique<ImplType>(std::move(Impl)));
+  OnDiskTrieRawHashMap::ImplType Impl{DatabaseFile(std::move(*File)), Trie};
+  return OnDiskTrieRawHashMap(std::make_unique<ImplType>(std::move(Impl)));
 }
 
 static Error createInvalidTrieError(uint64_t Offset, const Twine &Msg) {
@@ -1127,10 +838,15 @@ static Error createInvalidTrieError(uint64_t Offset, const Twine &Msg) {
 //===----------------------------------------------------------------------===//
 
 namespace {
-// A vistior to traverse the Trie.
+/// A multi-threaded vistior to traverse the Trie.
+///
+/// TODO: add more sanity checks that isn't just plain data corruption. For
+/// example, some ill-formed data can be constructed to form a cycle using
+/// Sub-Tries and it can lead to inifinite loop when visiting (or inserting
+/// data).
 class TrieVisitor {
 public:
-  TrieVisitor(HashMappedTrieHandle Trie, unsigned ThreadCount = 0,
+  TrieVisitor(TrieRawHashMapHandle Trie, unsigned ThreadCount = 0,
               unsigned ErrorLimit = 50)
       : Trie(Trie), ErrorLimit(ErrorLimit),
         Threads(hardware_concurrency(ThreadCount)) {}
@@ -1138,23 +854,26 @@ public:
   Error visit();
 
 private:
+  // Virtual method to implement the action when visiting a sub-trie.
   virtual Error visitSubTrie(StringRef Prefix, SubtrieHandle SubTrie) {
     return Error::success();
   }
 
+  // Virtual method to implement the action when visiting a slot in a trie node.
   virtual Error visitSlot(unsigned I, SubtrieHandle Subtrie, StringRef Prefix,
                           SubtrieSlotValue Slot) {
     return Error::success();
   }
 
 protected:
-  HashMappedTrieHandle Trie;
+  TrieRawHashMapHandle Trie;
 
 private:
   Error traverseTrieNode(SubtrieHandle Node, StringRef Prefix);
 
   Error validateSubTrie(SubtrieHandle Node, bool IsRoot);
 
+  // Helper function to capture errors when visiting the trie nodes.
   void addError(Error NewError) {
     assert(NewError && "not an error");
     std::lock_guard<std::mutex> ErrorLock(Lock);
@@ -1183,9 +902,10 @@ private:
   DefaultThreadPool Threads;
 };
 
+/// A visitor that traverse and print the Trie.
 class TriePrinter : public TrieVisitor {
 public:
-  TriePrinter(HashMappedTrieHandle Trie, raw_ostream &OS,
+  TriePrinter(TrieRawHashMapHandle Trie, raw_ostream &OS,
               function_ref<void(ArrayRef<char>)> PrintRecordData)
       : TrieVisitor(Trie, /*ThreadCount=*/1), OS(OS),
         PrintRecordData(PrintRecordData) {}
@@ -1197,7 +917,7 @@ public:
     OS << "records\n";
     llvm::sort(Records);
     for (int64_t Offset : Records) {
-      HashMappedTrieHandle::RecordData Record =
+      TrieRawHashMapHandle::RecordData Record =
           Trie.getRecord(SubtrieSlotValue::getDataOffset(Offset));
       if (auto Err = printRecord(Record))
         return Err;
@@ -1205,8 +925,8 @@ public:
     return Error::success();
   }
 
-  Error printRecord(HashMappedTrieHandle::RecordData &Record) {
-    OS << "- addr=" << (void*)Record.getFileOffset().get() << " ";
+  Error printRecord(TrieRawHashMapHandle::RecordData &Record) {
+    OS << "- addr=" << (void *)Record.getFileOffset().get() << " ";
     if (PrintRecordData) {
       PrintRecordData(Record.Proxy.Data);
     } else {
@@ -1249,7 +969,7 @@ public:
       OS << "\n";
       return Error::success();
     }
-    HashMappedTrieHandle::RecordData Record = Trie.getRecord(Slot);
+    TrieRawHashMapHandle::RecordData Record = Trie.getRecord(Slot);
     OS << "addr=" << (void *)Record.getFileOffset().get();
     OS << " content=";
     Subtrie.printHash(OS, Record.Proxy.Hash);
@@ -1264,12 +984,12 @@ private:
   SmallVector<int64_t> Records;
 };
 
-// TrieVerifier that adds additional verification on top of the basic visitor.
+/// TrieVerifier that adds additional verification on top of the basic visitor.
 class TrieVerifier : public TrieVisitor {
 public:
   TrieVerifier(
-      HashMappedTrieHandle Trie,
-      function_ref<Error(FileOffset, OnDiskHashMappedTrie::ConstValueProxy)>
+      TrieRawHashMapHandle Trie,
+      function_ref<Error(FileOffset, OnDiskTrieRawHashMap::ConstValueProxy)>
           RecordVerifier)
       : TrieVisitor(Trie), RecordVerifier(RecordVerifier) {}
 
@@ -1284,16 +1004,16 @@ private:
       if (!isAligned(MappedFileRegionArena::getAlign(), Slot.asData()))
         return createInvalidTrieError(Slot.asData(), "mis-aligned data entry");
 
-      HashMappedTrieHandle::RecordData Record =
+      TrieRawHashMapHandle::RecordData Record =
           Trie.getRecord(SubtrieSlotValue::getDataOffset(Slot.asData()));
       return RecordVerifier(Slot.asDataFileOffset(),
-                            OnDiskHashMappedTrie::ConstValueProxy{
+                            OnDiskTrieRawHashMap::ConstValueProxy{
                                 Record.Proxy.Hash, Record.Proxy.Data});
     }
     return Error::success();
   }
 
-  function_ref<Error(FileOffset, OnDiskHashMappedTrie::ConstValueProxy)>
+  function_ref<Error(FileOffset, OnDiskTrieRawHashMap::ConstValueProxy)>
       RecordVerifier;
 };
 } // namespace
@@ -1355,9 +1075,11 @@ Error TrieVisitor::validateSubTrie(SubtrieHandle Node, bool IsRoot) {
       Trie.getRegion().data() + Trie.getRegion().size())
     return createInvalidTrieError(Offset, "subtrie node spans out of bound");
 
-  if (Node.getStartBit() + Node.getNumBits() > Trie.getNumHashBits())
+  if (!IsRoot &&
+      Node.getStartBit() + Node.getNumBits() > Trie.getNumHashBits()) {
     return createInvalidTrieError(Offset,
                                   "subtrie represents too many hash bits");
+  }
 
   if (IsRoot) {
     if (Node.getStartBit() != 0)
@@ -1407,7 +1129,7 @@ Error TrieVisitor::traverseTrieNode(SubtrieHandle Node, StringRef Prefix) {
   return Error::success();
 }
 
-void HashMappedTrieHandle::print(
+void TrieRawHashMapHandle::print(
     raw_ostream &OS, function_ref<void(ArrayRef<char>)> PrintRecordData) const {
   OS << "hash-num-bits=" << getNumHashBits()
      << " hash-size=" << getNumHashBytes()
@@ -1423,8 +1145,8 @@ void HashMappedTrieHandle::print(
   return;
 }
 
-Error HashMappedTrieHandle::validate(
-    function_ref<Error(FileOffset, OnDiskHashMappedTrie::ConstValueProxy)>
+Error TrieRawHashMapHandle::validate(
+    function_ref<Error(FileOffset, OnDiskTrieRawHashMap::ConstValueProxy)>
         RecordVerifier) const {
   // Use the base TrieVisitor to identify the errors inside trie first.
   TrieVisitor BasicVerifier(*this);
@@ -1438,270 +1160,61 @@ Error HashMappedTrieHandle::validate(
   return Verifier.visit();
 }
 
-//===----------------------------------------------------------------------===//
-// DataAllocator data structures.
-//===----------------------------------------------------------------------===//
-
-namespace {
-/// DataAllocator table layout:
-/// - [8-bytes: Generic table header]
-/// - 8-bytes: AllocatorOffset (reserved for implementing free lists)
-/// - 8-bytes: Size for user data header
-/// - <user data buffer>
-///
-/// Record layout:
-/// - <data>
-class DataAllocatorHandle {
-public:
-  static constexpr TableHandle::TableKind Kind =
-      TableHandle::TableKind::DataAllocator;
-
-  struct Header {
-    TableHandle::Header GenericHeader;
-    std::atomic<int64_t> AllocatorOffset;
-    const uint64_t UserHeaderSize;
-  };
-
-  operator TableHandle() const {
-    if (!H)
-      return TableHandle();
-    return TableHandle(*Region, H->GenericHeader);
-  }
-
-  Expected<MutableArrayRef<char>> allocate(MappedFileRegionArena &Alloc,
-                                           size_t DataSize) {
-    assert(&Alloc.getRegion() == Region);
-    auto Ptr = Alloc.allocate(DataSize);
-    if (LLVM_UNLIKELY(!Ptr))
-      return Ptr.takeError();
-    return MutableArrayRef(*Ptr, DataSize);
-  }
-
-  explicit operator bool() const { return H; }
-  const Header &getHeader() const { return *H; }
-  MappedFileRegion &getRegion() const { return *Region; }
-
-  MutableArrayRef<uint8_t> getUserHeader() {
-    return MutableArrayRef(reinterpret_cast<uint8_t *>(H + 1),
-                           H->UserHeaderSize);
-  }
-
-  static Expected<DataAllocatorHandle>
-  create(MappedFileRegionArena &Alloc, StringRef Name, uint32_t UserHeaderSize);
-
-  DataAllocatorHandle() = default;
-  DataAllocatorHandle(MappedFileRegion &Region, Header &H)
-      : Region(&Region), H(&H) {}
-  DataAllocatorHandle(MappedFileRegion &Region, intptr_t HeaderOffset)
-      : DataAllocatorHandle(
-            Region, *reinterpret_cast<Header *>(Region.data() + HeaderOffset)) {
-  }
-
-private:
-  MappedFileRegion *Region = nullptr;
-  Header *H = nullptr;
-};
-
-} // end anonymous namespace
-
-struct OnDiskDataAllocator::ImplType {
-  DatabaseFile File;
-  DataAllocatorHandle Store;
-};
-
-Expected<DataAllocatorHandle>
-DataAllocatorHandle::create(MappedFileRegionArena &Alloc, StringRef Name,
-                            uint32_t UserHeaderSize) {
-  // Allocate.
-  auto Offset =
-      Alloc.allocateOffset(sizeof(Header) + UserHeaderSize + Name.size() + 1);
-  if (LLVM_UNLIKELY(!Offset))
-    return Offset.takeError();
-
-  // Construct the header and the name.
-  assert(Name.size() <= UINT16_MAX && "Expected smaller table name");
-  auto *H = new (Alloc.getRegion().data() + *Offset)
-      Header{{TableHandle::TableKind::DataAllocator, (uint16_t)Name.size(),
-              (int32_t)(sizeof(Header) + UserHeaderSize)},
-             /*AllocatorOffset=*/{0},
-             /*UserHeaderSize=*/UserHeaderSize};
-  memset(H + 1, 0, UserHeaderSize);
-  char *NameStorage = reinterpret_cast<char *>(H + 1) + UserHeaderSize;
-  llvm::copy(Name, NameStorage);
-  NameStorage[Name.size()] = 0;
-  return DataAllocatorHandle(Alloc.getRegion(), *H);
-}
-
-Expected<OnDiskDataAllocator> OnDiskDataAllocator::create(
-    const Twine &PathTwine, const Twine &TableNameTwine, uint64_t MaxFileSize,
-    std::optional<uint64_t> NewFileInitialSize, uint32_t UserHeaderSize,
-    std::shared_ptr<ondisk::OnDiskCASLogger> Logger,
-    function_ref<void(void *)> UserHeaderInit) {
-  assert(!UserHeaderSize || UserHeaderInit);
-  SmallString<128> PathStorage;
-  StringRef Path = PathTwine.toStringRef(PathStorage);
-  SmallString<128> TableNameStorage;
-  StringRef TableName = TableNameTwine.toStringRef(TableNameStorage);
-
-  // Constructor for if the file doesn't exist.
-  auto NewDBConstructor = [&](DatabaseFile &DB) -> Error {
-    auto Store =
-        DataAllocatorHandle::create(DB.getAlloc(), TableName, UserHeaderSize);
-    if (LLVM_UNLIKELY(!Store))
-      return Store.takeError();
-
-    DB.addTable(*Store);
-    if (UserHeaderSize)
-      UserHeaderInit(Store->getUserHeader().data());
-    return Error::success();
-  };
-
-  // Get or create the file.
-  Expected<DatabaseFile> File =
-      DatabaseFile::create(Path, MaxFileSize, Logger, NewDBConstructor);
-  if (!File)
-    return File.takeError();
-
-  // Find the table and validate it.
-  //
-  // TODO: Add support for creating/adding a table to an existing file.
-  std::optional<TableHandle> Table = File->findTable(TableName);
-  if (!Table)
-    return createTableConfigError(std::errc::argument_out_of_domain, Path,
-                                  TableName, "table not found");
-  if (Error E = checkTable("table kind", (size_t)DataAllocatorHandle::Kind,
-                           (size_t)Table->getHeader().Kind, Path, TableName))
-    return std::move(E);
-  auto Store = Table->cast<DataAllocatorHandle>();
-  assert(Store && "Already checked the kind");
-
-  // Success.
-  OnDiskDataAllocator::ImplType Impl{DatabaseFile(std::move(*File)), Store};
-  return OnDiskDataAllocator(std::make_unique<ImplType>(std::move(Impl)));
-}
-
-Expected<OnDiskDataAllocator::pointer>
-OnDiskDataAllocator::allocate(size_t Size) {
-  auto Data = Impl->Store.allocate(Impl->File.getAlloc(), Size);
-  if (LLVM_UNLIKELY(!Data))
-    return Data.takeError();
-
-  return pointer(FileOffset(Data->data() - Impl->Store.getRegion().data()),
-                 *Data);
-}
-
-const char *OnDiskDataAllocator::beginData(FileOffset Offset) const {
-  assert(Offset);
-  assert(Impl);
-  assert(Offset.get() < (int64_t)Impl->File.getAlloc().size());
-  return Impl->File.getRegion().data() + Offset.get();
-}
-
-MutableArrayRef<uint8_t> OnDiskDataAllocator::getUserHeader() {
-  return Impl->Store.getUserHeader();
-}
-
-size_t OnDiskDataAllocator::size() const { return Impl->File.size(); }
-size_t OnDiskDataAllocator::capacity() const {
-  return Impl->File.getRegion().size();
-}
-
-OnDiskDataAllocator::OnDiskDataAllocator(std::unique_ptr<ImplType> Impl)
-    : Impl(std::move(Impl)) {}
-
 #else // !LLVM_ENABLE_ONDISK_CAS
 
-struct OnDiskHashMappedTrie::ImplType {};
+struct OnDiskTrieRawHashMap::ImplType {};
 
-Expected<OnDiskHashMappedTrie>
-OnDiskHashMappedTrie::create(const Twine &PathTwine, const Twine &TrieNameTwine,
+Expected<OnDiskTrieRawHashMap>
+OnDiskTrieRawHashMap::create(const Twine &PathTwine, const Twine &TrieNameTwine,
                              size_t NumHashBits, uint64_t DataSize,
                              uint64_t MaxFileSize,
                              std::optional<uint64_t> NewFileInitialSize,
                              std::shared_ptr<OnDiskCASLogger> Logger,
                              std::optional<size_t> NewTableNumRootBits,
                              std::optional<size_t> NewTableNumSubtrieBits) {
-  report_fatal_error("not supported");
+  return createStringError(make_error_code(std::errc::not_supported),
+                           "OnDiskTrieRawHashMap is not supported");
 }
 
-Expected<OnDiskHashMappedTrie::pointer>
-OnDiskHashMappedTrie::insertLazy(ArrayRef<uint8_t> Hash,
+Expected<OnDiskTrieRawHashMap::pointer>
+OnDiskTrieRawHashMap::insertLazy(ArrayRef<uint8_t> Hash,
                                  LazyInsertOnConstructCB OnConstruct,
                                  LazyInsertOnLeakCB OnLeak) {
-  report_fatal_error("not supported");
+  return createStringError(make_error_code(std::errc::not_supported),
+                           "OnDiskTrieRawHashMap is not supported");
 }
 
-OnDiskHashMappedTrie::const_pointer
-OnDiskHashMappedTrie::recoverFromFileOffset(FileOffset Offset) const {
-  report_fatal_error("not supported");
+Expected<OnDiskTrieRawHashMap::const_pointer>
+OnDiskTrieRawHashMap::recoverFromFileOffset(FileOffset Offset) const {
+  return createStringError(make_error_code(std::errc::not_supported),
+                           "OnDiskTrieRawHashMap is not supported");
 }
 
-OnDiskHashMappedTrie::const_pointer
-OnDiskHashMappedTrie::find(ArrayRef<uint8_t> Hash) const {
-  report_fatal_error("not supported");
+OnDiskTrieRawHashMap::const_pointer
+OnDiskTrieRawHashMap::find(ArrayRef<uint8_t> Hash) const {
+  return const_pointer();
 }
 
-void OnDiskHashMappedTrie::print(
+void OnDiskTrieRawHashMap::print(
     raw_ostream &OS, function_ref<void(ArrayRef<char>)> PrintRecordData) const {
-  report_fatal_error("not supported");
 }
 
-Error OnDiskHashMappedTrie::validate(
-    function_ref<Error(FileOffset, OnDiskHashMappedTrie::ConstValueProxy)>
+Error OnDiskTrieRawHashMap::validate(
+    function_ref<Error(FileOffset, OnDiskTrieRawHashMap::ConstValueProxy)>
         RecordVerifier) const {
-  report_fatal_error("not supported");
+  return createStringError(make_error_code(std::errc::not_supported),
+                           "OnDiskTrieRawHashMap is not supported");
 }
 
-size_t OnDiskHashMappedTrie::size() const {
-  report_fatal_error("not supported");
-}
-
-size_t OnDiskHashMappedTrie::capacity() const {
-  report_fatal_error("not supported");
-}
-
-size_t OnDiskDataAllocator::capacity() const {
-  report_fatal_error("not supported");
-}
-
-struct OnDiskDataAllocator::ImplType {};
-
-Expected<OnDiskDataAllocator> OnDiskDataAllocator::create(
-    const Twine &Path, const Twine &TableName, uint64_t MaxFileSize,
-    std::optional<uint64_t> NewFileInitialSize, uint32_t UserHeaderSize,
-    std::shared_ptr<OnDiskCASLogger> Logger,
-    function_ref<void(void *)> UserHeaderInit) {
-  report_fatal_error("not supported");
-}
-
-Expected<OnDiskDataAllocator::pointer>
-OnDiskDataAllocator::allocate(size_t Size) {
-  report_fatal_error("not supported");
-}
-
-const char *OnDiskDataAllocator::beginData(FileOffset Offset) const {
-  report_fatal_error("not supported");
-}
-
-MutableArrayRef<uint8_t> OnDiskDataAllocator::getUserHeader() {
-  report_fatal_error("not supported");
-}
-
-size_t OnDiskDataAllocator::size() const {
-  report_fatal_error("not supported");
-}
+size_t OnDiskTrieRawHashMap::size() const { return 0; }
+size_t OnDiskTrieRawHashMap::capacity() const { return 0; }
 
 #endif // LLVM_ENABLE_ONDISK_CAS
 
-OnDiskHashMappedTrie::OnDiskHashMappedTrie(std::unique_ptr<ImplType> Impl)
+OnDiskTrieRawHashMap::OnDiskTrieRawHashMap(std::unique_ptr<ImplType> Impl)
     : Impl(std::move(Impl)) {}
-OnDiskHashMappedTrie::OnDiskHashMappedTrie(OnDiskHashMappedTrie &&RHS) =
+OnDiskTrieRawHashMap::OnDiskTrieRawHashMap(OnDiskTrieRawHashMap &&RHS) =
     default;
-OnDiskHashMappedTrie &
-OnDiskHashMappedTrie::operator=(OnDiskHashMappedTrie &&RHS) = default;
-OnDiskHashMappedTrie::~OnDiskHashMappedTrie() = default;
-
-OnDiskDataAllocator::OnDiskDataAllocator(OnDiskDataAllocator &&RHS) = default;
-OnDiskDataAllocator &
-OnDiskDataAllocator::operator=(OnDiskDataAllocator &&RHS) = default;
-OnDiskDataAllocator::~OnDiskDataAllocator() = default;
+OnDiskTrieRawHashMap &
+OnDiskTrieRawHashMap::operator=(OnDiskTrieRawHashMap &&RHS) = default;
+OnDiskTrieRawHashMap::~OnDiskTrieRawHashMap() = default;
